Move WithTooltip to components
